optimize code: move flat_list to osc2_helper
